Improve team support for WorkEffort and Request.
Teams and parties are both assigned via the regular Party joining
tables; if calling code wants to then check for membership, an OR
condition should be used, and test for both currentParty and against
teamMemberPartyId.

Additionally, if checking against the team, be sure to add a date filter
on teamRelationshipFromPartyId/thruPartyId.